refactor: reorganize hooks folder
